% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{anova.zibellreg}
\alias{anova.zibellreg}
\title{anova method for zibellreg models}
\usage{
\method{anova}{zibellreg}(...)
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\value{
the ANOVA table.
}
\description{
Compute analysis of variance (or deviance) tables for one or more fitted model objects.
}
\examples{
\donttest{
library(bellreg)
data(cells)
fit1 <- zibellreg(cells ~ 1|smoker+gender, data = cells, approach = "mle")
fit2 <- zibellreg(cells ~ smoker+gender|smoker+gender, data = cells, approach = "mle")
anova(fit1, fit2)
}

}
