% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellreg-package.R
\docType{package}
\name{bellreg-package}
\alias{bellreg-package}
\alias{bellreg}
\title{The 'bellreg' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.5. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fndemarqui/bellreg}
  \item \url{https://fndemarqui.github.io/bellreg/}
  \item Report bugs at \url{https://github.com/fndemarqui/bellreg/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
