% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.bellreg}
\alias{coef.bellreg}
\alias{coef.zibellreg}
\title{Estimated regression coefficients}
\usage{
\method{coef}{bellreg}(object, ...)

\method{coef}{zibellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated regression coefficients

the estimated regression coefficients
}
\description{
Estimated regression coefficients

Estimated regression coefficients
}
