% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.bellreg}
\alias{coef.bellreg}
\title{Estimated regression coefficients for the bellreg model}
\usage{
\method{coef}{bellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector with the estimated regression coefficients.
}
\description{
Estimated regression coefficients for the bellreg model
}
\examples{
\donttest{
fit <- bellreg(nf ~ lroll, data=faults)
coef(fit)
}

}
