% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coef.zibellreg}
\alias{coef.zibellreg}
\title{Estimated regression coefficients for zibellreg model}
\usage{
\method{coef}{zibellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list containing the the estimated regression coefficients associated with the degenerated and Bell count distributions, respectively.
}
\description{
Estimated regression coefficients for zibellreg model
}
\examples{
\donttest{
fit <- zibellreg(cells ~ smoker + gender|smoker + gender, data = cells)
coef(fit)
}

}
