% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmeans.R
\name{emmeans-bellreg-helpers}
\alias{emmeans-bellreg-helpers}
\alias{recover_data.bellreg}
\alias{recover_data.zibellreg}
\title{Support Functions for \pkg{emmeans}}
\usage{
\method{recover_data}{bellreg}(object, ...)

\method{recover_data}{zibellreg}(object, term = c("zero", "count"), ...)
}
\arguments{
\item{object}{An object of the same class as is supported by a new method.}

\item{...}{Additional parameters that may be supported by the method.}

\item{term}{character specifying whether zero or count term regression coefficients are to be used.}
}
\description{
Functions required for compatibility of \pkg{bellreg} with \pkg{emmeans}.
Users are not required to call these functions themselves. Instead,
they will be called automatically by the \code{emmeans} function
of the \pkg{emmeans} package.
}
