% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{faults}
\alias{faults}
\title{Faults data set}
\format{
A data frame with 32 rows and 2 variables:
\itemize{
\item nf: number of faults in rolls of fabric of different lengths.
\item lroll: length of the roll.
}
}
\description{
Data set taken from \insertCite{1982_Hinde;textual}{bellreg} and posteriorly analyzed by \insertCite{2018_Castellares;textual}{bellreg}. The data contains the number of faults in rolls of fabric of different lengths.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
