% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{fitted.bellreg}
\alias{fitted.bellreg}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{bellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a vector with the fitted values (for MLE approach) or a matrix containing the posterior sample of the fitted values.
}
\description{
This function returns the fitted values.
}
\examples{
\donttest{
data(faults)
fit <- bellreg(nf ~ lroll, data = faults)
fitted.values(fit)
}

}
