% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.bellreg}
\alias{logLik.bellreg}
\title{Extract Log-Likelihood from a Fitted Model}
\usage{
\method{logLik}{bellreg}(object, ...)
}
\arguments{
\item{object}{a fitted model of the class bellreg}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the log-likelihood value when a single model is passed to the function; otherwise, a data.frame with the log-likelihood values and the number of parameters is returned.
}
\description{
Extracts the log-likelihood function for a fitted parametric model.
}
\examples{
\donttest{
library(bellreg)
data(faults)
fit1 <- bellreg(nf ~ 0 + lroll, data = faults, approach = "mle")
fit2 <- bellreg(nf ~ lroll, data = faults, approach = "mle")
logLik(fit1, fit2)
}

}
