% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.zibellreg}
\alias{logLik.zibellreg}
\title{Extract Log-Likelihood from a Fitted Model}
\usage{
\method{logLik}{zibellreg}(object, ...)
}
\arguments{
\item{object}{a fitted model of the class zibellreg}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the log-likelihood value when a single model is passed to the function; otherwise, a data.frame with the log-likelihood values and the number of parameters is returned.
}
\description{
Extracts the log-likelihood function for a fitted parametric model.
}
\examples{
\donttest{
library(bellreg)
data(cells)
fit1 <- zibellreg(cells ~ 1|smoker+gender, data = cells, approach = "mle")
fit2 <- zibellreg(cells ~ smoker+gender|smoker+gender, data = cells, approach = "mle")
logLik(fit1, fit2)
}

}
