% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix.R
\name{model.matrix.zibellreg}
\alias{model.matrix.zibellreg}
\title{Construct Design Matrices for zibellreg models}
\usage{
\method{model.matrix}{zibellreg}(object, which = c("all", "zero", "count"), ...)
}
\arguments{
\item{object}{an object of class 'zibellreg'.}

\item{which}{character indicating which design matrix to extract: "all" (default), "zero", or "count".}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A design matrix.
}
\description{
Construct Design Matrices for zibellreg models
}
\examples{
library(bellreg)
data(cells)
fit <- zibellreg(formula = cells ~ smoker + gender | age, data = cells, approach = "mle")
head(model.matrix(fit))
head(model.matrix(fit, "all"))
head(model.matrix(fit, "zero"))
head(model.matrix(fit, "count"))

}
