% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bellreg.R
\name{print.summary.bellreg}
\alias{print.summary.bellreg}
\title{Print the summary.bellreg output}
\usage{
\method{print}{summary.bellreg}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.bellreg.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.bellreg output
}
