% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_zibellreg.R
\name{print.summary.zibellreg}
\alias{print.summary.zibellreg}
\title{Print the summary.zibellreg output}
\usage{
\method{print}{summary.zibellreg}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.zibellreg.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.zibellreg output
}
