% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{prior_spec}
\alias{prior_spec}
\title{Prior specification}
\usage{
prior_spec(
  dists = list(intercept ~ normal(0, 10), beta ~ normal(0, 2.5)),
  autoscale = TRUE
)
}
\arguments{
\item{dists}{a list containing the prior specification for the parameters.}

\item{autoscale}{logical argument indicating if the covariates should be centered and scaled (default is TRUE)}
}
\description{
Prior specification
}
