% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_bellreg.R
\name{summary.bellreg}
\alias{summary.bellreg}
\title{Summary for the bellreg model}
\usage{
\method{summary}{bellreg}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'bellreg'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the bellreg model
}
