% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_zibellreg.R
\name{summary.zibellreg}
\alias{summary.zibellreg}
\title{Summary for the zibellreg model}
\usage{
\method{summary}{zibellreg}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'zibellreg'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the zibellreg model
}
