% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy.zibellreg}
\alias{tidy.zibellreg}
\title{Tidy a zibellreg object}
\usage{
\method{tidy}{zibellreg}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a fitted model object.}

\item{conf.int}{Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{the confidence level required.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a tibble with a summary of the fit.
}
\description{
Tidy a zibellreg object
}
\details{
Convert a fitted model into a tibble.
}
\examples{
\donttest{
library(bellreg)
data(cells)
fit <- zibellreg(cells ~ smoker+gender|smoker+gender, data = cells, approach = "mle")
tidy(fit)
}

}
