% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.bellreg}
\alias{vcov.bellreg}
\title{Covariance of the regression coefficients}
\usage{
\method{vcov}{bellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
Covariance of the regression coefficients
}
