% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcov.zibellreg}
\alias{vcov.zibellreg}
\title{Covariance of the regression coefficients}
\usage{
\method{vcov}{zibellreg}(object, ...)
}
\arguments{
\item{object}{an object of the class bellreg}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
Covariance of the regression coefficients
}
\examples{
\donttest{
fit <- zibellreg(cells ~ smoker + gender|smoker + gender, data = cells)
vcov(fit)
}

}
