% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zibellreg.R
\name{zibellreg}
\alias{zibellreg}
\title{ZiBell regression model}
\usage{
zibellreg(
  formula,
  data,
  approach = c("mle", "bayes"),
  hessian = TRUE,
  link1 = c("logit", "probit", "cloglog", "cauchy"),
  link2 = c("log", "sqrt", "identity"),
  priors = prior_spec(list(intercept ~ normal(0, 10), beta ~ normal(0, 2.5)), autoscale =
    TRUE),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which ypbp is called.}

\item{approach}{approach to be used to fit the model (mle: maximum likelihood; bayes: Bayesian approach).}

\item{hessian}{hessian logical; If TRUE (default), the hessian matrix is returned when approach="mle".}

\item{link1}{assumed link function for degenerate distribution (logit, probit, cloglog, cauchy); default is logit.}

\item{link2}{assumed link function for count distribution (log, sqrt or identiy); default is log.}

\item{priors}{a list containing the prior specification for the parameters; if NULL, default prior are used.}

\item{...}{further arguments passed to either \code{rstan::optimizing} or \code{rstan::sampling}.}
}
\value{
zibellreg returns an object of class "zibellreg" containing the fitted model.
}
\description{
Fits the Bell regression model to overdispersed count data.
}
\examples{
\donttest{
# ML approach:
data(cells)
mle <- zibellreg(cells ~ smoker+gender|smoker+gender, data = cells, approach = "mle")
summary(mle)

# Bayesian approach:
bayes <- zibellreg(cells ~ 1|smoker+gender, data = cells, approach = "bayes", refresh = FALSE)
summary(bayes)
}

}
