% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{AIC.bellreg}
\alias{AIC.bellreg}
\title{Akaike Information Criterion for the bellreg model}
\usage{
\method{AIC}{bellreg}(object, ..., k = 2)
}
\arguments{
\item{object}{an object of the class bellreg.}

\item{...}{further arguments passed to or from other methods.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\value{
The AIC associated with the fitted model.
}
\description{
Function to compute the Akaike Information Criterion (AIC) for fitted bellreg model.
}
\examples{
\dontrun{
fit <- bellreg(nf~lroll, data=faults, init=0)
AIC(fit)
}

}
