% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bellreg-package.R
\docType{package}
\name{bellreg-package}
\alias{bellreg-package}
\alias{BellReg}
\title{The 'bellreg' package.}
\description{
Bell Regression models for count data with overdispersion. The implemented models account for ordinary and zero-inflated regression models under both frequentist and Bayesian approaches. Theorical details regarding the models implemented in the package can be found in \insertCite{2018_Castellares}{bellreg} and \insertCite{2020_Lemonte}{bellreg}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org

\insertRef{2018_Castellares}{bellreg}

\insertRef{2020_Lemonte}{bellreg}
}
