% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.bellreg}
\alias{confint.bellreg}
\title{Confidence intervals for the regression coefficients}
\usage{
\method{confint}{bellreg}(object, level = 0.95, ...)
}
\arguments{
\item{object}{an object of the class bellreg}

\item{level}{the confidence level required}

\item{...}{further arguments passed to or from other methods}
}
\value{
100(1-alpha)% confidence intervals for the regression coefficients
}
\description{
Confidence intervals for the regression coefficients
}
\examples{
\donttest{
fit <- bellreg(nf ~ lroll, data = faults)
confint(fit)
}

}
